Beaker.Router.Helpers

Module with named helpers generated from Beaker.Router.

metrics_path(conn_or_endpoint, atom2)

metrics_path(conn_or_endpoint, atom2, params)

metrics_url(conn_or_endpoint, atom2)

metrics_url(conn_or_endpoint, atom2, params)

path(data, path)

Generates the path information including any necessary prefix

static_path(conn, path)

Generates path to a static asset given its file path

static_url(conn, path)

Generates url to a static asset given its file path

url(data)

Generates the connection/endpoint base URL without any path information

Functions

metrics_path(conn_or_endpoint, atom2)

metrics_path(conn_or_endpoint, atom2, params)

metrics_url(conn_or_endpoint, atom2)

metrics_url(conn_or_endpoint, atom2, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.