beaker v1.2.0 Beaker.Router.Helpers

Module with named helpers generated from Beaker.Router.

Functions

metrics_api_path(conn_or_endpoint, action)

metrics_api_path(conn_or_endpoint, action, params)

metrics_api_url(conn_or_endpoint, action)

metrics_api_url(conn_or_endpoint, action, params)

metrics_path(conn_or_endpoint, action)

metrics_path(conn_or_endpoint, action, params)

metrics_url(conn_or_endpoint, action)

metrics_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.