beaker v1.2.0 Beaker.TimeSeries.Aggregator

Beaker.TimeSeries.Aggregator is the aggregation framework for time series. It’s purposed mostly for internal use only. Currently, it’s not configurable, and set to run every 60 seconds. This will change in the future to allow for more granularity.

The only function that developers should use is last_aggregated_at/0 to check when aggregation was last run.

Functions

last_aggregated_at()

Get the last time aggregation was run for time series.

Examples

iex> Beaker.TimeSeries.Aggregator.last_aggregated_at 1442946540000000

Returns last_aggregated_at where last_aggregated_at is a timestamp of when the last time aggregation was successfully run. The timestamp is in epoch style, and in microseconds.